requirements.txt
torch>=2.0.0
numpy>=1.24.0
pillow>=9.0.0
transformers>=4.40.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
